services/client_portal_service/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-filter>=24.0,<25.0
mysqlclient>=2.2,<3.0
drf-spectacular>=0.27,<1.0
gunicorn>=22.0,<23.0
PyJWT>=2.8,<3.0
requests>=2.31,<3.0
